+++ src/components/AlertMessages/AlertMessagesProvider.tsx
@@ -1,8 +1,5 @@
 import { createContext, ReactNode, useContext, useState } from 'react';
-
-export const ALERT_COLORS = ['info', 'success', 'warning', 'error'] as const;
-
-export type AlertColor = (typeof ALERT_COLORS)[number];
+import { AlertColor } from '../Alert';
 
 export interface AlertMessage {
   color?: AlertColor;
@@ -22,7 +19,7 @@ export interface AlertMessagesProviderProps {
   initialData?: AlertMessage[];
 }
 
-const initialContext: AlertMessagesContextData = {
+const initialContext = {
   alertMessages: [],
   addAlertMessages: () => undefined,
   clearAlertMessages: () => undefined,
@@ -32,13 +29,12 @@ const initialContext: AlertMessagesContextData = {
 const AlertMessagesContext =
   createContext<AlertMessagesContextData>(initialContext);
 
-export const useAlertMessages = (): AlertMessagesContextData =>
-  useContext(AlertMessagesContext);
+export const useAlertMessages = () => useContext(AlertMessagesContext);
 
 export const AlertMessagesProvider = ({
   children,
   initialData,
-}: AlertMessagesProviderProps): JSX.Element => {
+}: AlertMessagesProviderProps) => {
   const [alertMessages, setAlertMessages] = useState<AlertMessage[]>(
     initialData ?? initialContext.alertMessages,
   );