+++ src/components/Button/Button.stories.tsx
@@ -24,9 +24,58 @@ Loading.args = {
   loading: true,
 };
 
-export const Large: Story<ButtonProps> = args => <Button {...args} />;
+export const CustomColor: Story<ButtonProps> = args => (
+  <div className="flex gap-2">
+    <Button {...args}>Default</Button>
+    <Button color="neutral" {...args}>
+      Neutral
+    </Button>
+    <Button color="ghost" {...args}>
+      Ghost
+    </Button>
+    <Button color="primary" {...args}>
+      Primary
+    </Button>
+    <Button color="secondary" {...args}>
+      Secondary
+    </Button>
+    <Button color="accent" {...args}>
+      Accent
+    </Button>
+    <Button color="info" {...args}>
+      Info
+    </Button>
+    <Button color="success" {...args}>
+      Success
+    </Button>
+    <Button color="warning" {...args}>
+      Warning
+    </Button>
+    <Button color="error" {...args}>
+      Error
+    </Button>
+  </div>
+);
 
-Large.args = {
-  children: 'Large',
-  size: 'lg',
+CustomColor.args = {
+  size: 'md',
 };
+
+export const CustomSize: Story<ButtonProps> = args => (
+  <div className="flex items-center gap-2">
+    <Button size="xs" {...args}>
+      Extra Small
+    </Button>
+    <Button size="sm" {...args}>
+      Small
+    </Button>
+    <Button size="md" {...args}>
+      Medium
+    </Button>
+    <Button size="lg" {...args}>
+      Large
+    </Button>
+  </div>
+);
+
+CustomSize.args = {};