+++ src/components/Card/CardActions.tsx
@@ -1,0 +1,11 @@
+import classNames from 'classnames';
+import { HTMLAttributes } from 'react';
+
+export type CardActionsProps = HTMLAttributes<HTMLDivElement>;
+
+export const CardActions = ({ className, ...props }: CardActionsProps) => (
+  <div
+    className={classNames('card-actions justify-end', className)}
+    {...props}
+  />
+);