+++ src/components/FullScreenLoader/FullScreenLoader.tsx
@@ -1,4 +1,4 @@
-export const FullScreenLoader = (): JSX.Element => (
+export const FullScreenLoader = () => (
   <div className="flex flex-1 items-center justify-center">
     <progress className="progress w-56" />
   </div>