+++ src/components/Stat/StatDesc.tsx
@@ -1,0 +1,8 @@
+import classNames from 'classnames';
+import { HTMLAttributes } from 'react';
+
+export type StatDescProps = HTMLAttributes<HTMLDivElement>;
+
+export const StatDesc = ({ className, ...props }: StatDescProps) => (
+  <div className={classNames('stat-desc', className)} {...props} />
+);